+++ include/runtime.h
@@ -1,0 +1,7 @@
+#pragma once
+
+void applyRuntimeSettings(bool redrawNow);
+void processPendingSystemActions();
+void requestWifiReset();
+void requestSettingsReset();
+void requestFactoryReset();